DIFFCALC: Derived and added a new fixed mu=0, chi=90, psi mode and test case for I07's vertical surface normal geometry.
The tests currebtly fail as the values differ from those predicted by Willmotts engine.
The difference in gamma/phi from willmotts's number is due to his choice to flip X
(x component of q in the lab frame). There is no trivial way of applying the same reflection
(or indeed to get the same effect of making gamma>0) in diffcalc that I can see now.

When reflections have been found at the predicted values (with -ve gammas) the tests will be
modified to reflect this and should then pass.

This still leaves us needing a way to flip q_lab_X.